:fire: MergeConflict objects
These are no longer necessary since we've simplified StagingView models
to be plain ole JS objects. And for now we're simply displaying the
merge conflict file as it appears in the working directory. In the
future we may want to display a special UI for resolving merge
conflicts, much like Ash Wilson's
https://atom.io/packages/merge-conflicts