Replace EtchReactAdapter function with EtchWrapper component
Previously, rendering an Etch component with React required creating a
new class by passing the Etch component class to `EtchReactAdapter` and
rendering what was returned with React. Instead, EtchReactAdapter allows
you to specify which component to wrap and what props to pass it
declaratively via JSX, and will manage the component lifecycle itself.