Fix bad logic in GithubPackage#repositoryForProjectDirectory
If this function was called multiple times before the first call to
Repository.open resolves, it ended up clobbering the cache, resulting in
multiple Repository objects returned for the same working directory.

Additionally, it depended on referential equality of the passed
Directory object instead of the path of that Directory, causing
different repos to be returned for Directories with the same working
directory path.