Move GSOS#isGitRepository to GSOS#resolveDotGitDir
This allows us to use the operation we were running to determine if a
workdir belonged to a valid repository to also resolve the actual path
of the .git directory (in the case of a `gitdir:` link).

When a repository determines whether or not it actually exists, it also
caches the path to its .git directory for future use.

Any methods in GSOS that needed to access files in the .git directory
now have the path to the directory passed in.