Fix stale function reference when calling context menu handler
Upon mounting, ContextMenuInterceptor retained the onWillShowContextMenu
prop as a direct function reference; when updating the component with
a new onWillShowContextMenu, the retained function was not updated. As
a result, right-clicking on the wrapped element could potentially call
a previously-passed value of onWillShowContextMenu. The solution is to
instead retain a function that has dynamic access to the latest
onWillShowContextMenu prop.