Handle viewing changes for file not in repo
Notify user that repo does not exist, and show button to initialize
repo. If the user initializes the repo, we retry the original operation
to view changes for the opened file.

When showing the init dialog, the initial path is set to the project
folder that contains the active text editor, if any.