Scroll to selected diff line only if head of selection changes
Previously, `componentDidUpdate` was being called when anonymous
function props were updated, triggering scrolling at undesired times.
Instead, now we specifically check to see if the relevant props have
changed before scrolling.

Admittedly, I'm not sure what the purpose of the `selectedLine` branch
of the original conditional was for. I traced its addition back to the
"open diff from editor" feature I added 8 months ago, but manual testing
of that feature seems to work without it. So I'm removing it and hoping
that no regression is introduced, and taking a mental note to write more
informative commit messages in the future.