PR feedback: Use Tooltip, name action (push-/pull-/fetching), animations
This commit refactors the component to not use atom.tooltips but the
internal <Tooltip> component instead.
It displays "Pushing"/"Pulling"/"Fetching" in the status bar tile when
such an action is in progress.
It refactors the component to extract all possible tile states into a
tileState object.
It adds animations to the "Pushing" and "Pulling" status bar icons.