fix issue with setting cursor buffer position
This only works for commit message templates where every line starts 
with a #.  We explored using isCommitMessageClean to try and handle 
commit message templates that have uncommented lines.  We're not calling 
didUpdate every time the commit message is updated, so it's out of sync, 
and we decided it was getting too edge casey out there.

Co-Authored-By: Katrina Uychaco <katrina@github.com>