Pass {before, after} Marker sets to FilePatch::triggerCollapseIn()
This allows the PatchBuffer::extractPatchBuffer() call within the
collapse operation to correctly identify which boundary markers should
be included in the extracted PatchBuffer and which should not, even when
the FilePatch's marker is itself empty.

I've also refactored out MultiFilePatch::getMarkersBefore() and
MultiFilePatch::getMarkersAfter() methods to save some duplication among
triggerExpandIn() and triggerCollapseIn() methods, and added a test case
that minimally reproduces the lost marker.