Refactor a createSubBuffer() method out of extractPatchBuffer()
For the mini-editors that show local diff context for each review
comment, we need to be able to _non-destructively_ extract a range from
a PatchBuffer, including markers and layers. extractPatchBuffer()
already did that but also removed the text and markers from the original
buffer, which is not what we want.