PatchBuffer::createSubBuffer() "slices" markers that cross its range
Previously, we didn't need extracted sub-buffers to worry about markers
that crossed the boundaries of the Range we were extracting, because we
were always collapsing entire FilePatch markers at once and none of the
markers we created could cross a FilePatch boundary. To support
"mini-diffs," though, we need to be able to create slices that could
cross regions (like only including one added row from a hunk, for
example). Fortunately the logic change wasn't too difficult.