Add fetchParams prop on ModelObserver
The ModelObserver component accepts a `fetchParams` property containing
additional arguments to pass to its `fetchData` function. Additionally,
an extra `fetchData` call will be performed if any of the `fetchParams`
change `===` identity.

@annthurium hit the need for this yesterday while working in the
CommentDecorationsContainer. The OAuth token fetched from a
GithubLoginModel depends on the Endpoint derived from the Repository's
remotes, but a change in the Endpoint did not trigger a re-fetch of the
token.

With this in place they should be able to do something like:

```jsx
class CommentDecorationsContainer extends React.Component {
  // ...

  renderWithRepositoryData = repoData => {
    const endpoint = repoData.currentRemote.getEndpoint();

    return (
      <ObserveModel
        model={this.props.loginModel}
        fetchParams={[endpoint.getLoginAccount()]}
        fetchData={this.fetchToken}>
        {token => this.renderWithToken(token, repoData)}
      </ObserveModel>
    )
  }

  fetchToken = (loginModel, account) => loginModel.getToken(account);

  renderWithToken(token, repoData) {
    // ...
  }
}
```