Extract CommentPositioning "info" objects into a class
Some refactoring to organize the "info" objects and make their structure
a bit clearer. Use a synthetic "filesystem event" to trigger the
Repository cache invalidation rather than a bespoke method, so we're
consistent with the way we handle changes done externally. Add a test
to cover the "update for file" flow.