Don't modify an object that is passed to components as props
React (and Relay) discourage you from modifying objects that are passed 
to components as props: 
https://reactjs.org/docs/components-and-props.html#props-are-read-only. 
The "data" passed to the Issueish constructor is sent directly from the 
props passed to IssueishListController, so we should treat them as 
immutable.

I thought that React froze objects passed as props in development mode 
to catch these issues. That's what's failing on CI - I have no idea why 
it isn't failing locally for us.