Add `NODE_ENV=development` to `test` script to see React errors
React runs fewer checks in production mode for efficiency. For example, in development mode it checks that props have not been mutated. We had an issue where tests were failing on CI due to mutated props but not locally until we set the NODE_ENV to be development. Let's just have this always be the case to err on the side of catching more errors locally.

Co-Authored-By: Ash Wilson <smashwilson@gmail.com>